Fix the size of a room_code

+++ src/game_api/level_api.hpp
@@ -17,7 +17,7 @@ struct TileCodeDef
 };
 struct TemplateDef
 {
-    std::uint32_t id;
+    std::uint16_t id;
 };
 struct ChanceDef
 {
